complete conversion to VirtueMart
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/trunk/virtuemart@20 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ html/product.file_list.php
@@ -5,7 +5,7 @@ defined( '_VALID_MOS' ) or die( 'Direct Access to this location is not allowed.'
 * @author Soeren Eberhardt
 * @param int product_id
 *
-* @version $Id: COPYRIGHT.php 70 2005-09-15 20:45:51Z spacemonkey $
+* @version $Id: product.file_list.php,v 1.3 2005/09/27 17:51:26 soeren_nb Exp $
 * @package VirtueMart
 * @subpackage html
 * @copyright Copyright (C) 2004-2005 Soeren Eberhardt. All rights reserved.
@@ -14,7 +14,7 @@ defined( '_VALID_MOS' ) or die( 'Direct Access to this location is not allowed.'
 * to the GNU General Public License, and as distributed it includes or
 * is derivative of works licensed under the GNU General Public License or
 * other free or open source software licenses.
-* See /administrator/components/com_phpshop/COPYRIGHT.php for copyright notices and details.
+* See /administrator/components/com_virtuemart/COPYRIGHT.php for copyright notices and details.
 *
 * http://virtuemart.net
 *
@@ -27,18 +27,23 @@ require_once( CLASSPATH . "htmlTools.class.php" );
 $product_id = mosGetParam($_REQUEST, 'product_id' );
 $task = mosGetParam($_REQUEST, 'task' );
 
-$q  = "SELECT product_name FROM #__pshop_product WHERE product_id = '$product_id'";
+$q  = "SELECT product_name FROM #__{vm}_product WHERE product_id = '$product_id'";
 $db->query($q);
 $db->next_record();
 $product_name = $db->f("product_name");
 
-$q = "SELECT file_id, file_is_image, file_product_id, file_extension, file_url, file_published, file_name, file_title FROM #__pshop_product_files  ";
+$dbf = new ps_DB;
+$sql = 'SELECT attribute_value FROM #__{vm}_product_attribute WHERE `product_id` = \''.$product_id.'\' AND attribute_name=\'download\'';
+$dbf->query( $sql );
+$dbf->next_record();
+	
+$q = "SELECT file_id, file_is_image, file_product_id, file_extension, file_url, file_published, file_name, file_title FROM #__{vm}_product_files  ";
 $q .= "WHERE file_product_id = '$product_id' ";
 $q .= "ORDER BY file_is_image ";
 $db->query($q);
 $db->next_record();
 if( $db->num_rows() < 1 && $task != "cancel" ) {
-  mosRedirect( $_SERVER['PHP_SELF']."?option=com_phpshop&page=product.file_form&product_id=$product_id" );
+  mosRedirect( $_SERVER['PHP_SELF']."?option=com_virtuemart&page=product.file_form&product_id=$product_id" );
 }
 else
 	$num_rows = $db->num_rows();
@@ -50,7 +55,7 @@ $pageNav = new vmPageNav( $num_rows, $limitstart, $limit );
 $listObj = new listFactory( $pageNav );
 
 // print out the search field and a list heading
-$listObj->writeSearchHeader($PHPSHOP_LANG->_PHPSHOP_FILES_LIST ." " . $product_name, $mosConfig_live_site."/administrator/images/mediamanager.png", $modulename, "file_list");
+$listObj->writeSearchHeader($VM_LANG->_PHPSHOP_FILES_LIST ." " . $product_name, $mosConfig_live_site."/administrator/images/mediamanager.png", $modulename, "file_list");
 
 // start the list table
 $listObj->startTable();
@@ -58,12 +63,12 @@ $listObj->startTable();
 // these are the columns in the table
 $columns = Array(  "#" => 'width="20"', 
 					"<input type=\"checkbox\" name=\"toggle\" value=\"\" onclick=\"checkAll($num_rows)\" />" => 'width="20"',
-					$PHPSHOP_LANG->_PHPSHOP_FILES_LIST_FILENAME => '',
-					$PHPSHOP_LANG->_PHPSHOP_VIEW => '',
-					$PHPSHOP_LANG->_PHPSHOP_FILES_LIST_FILETITLE => '',
-					$PHPSHOP_LANG->_PHPSHOP_UPDATE => '',
-					$PHPSHOP_LANG->_PHPSHOP_FILES_LIST_FILETYPE => '',
-					$PHPSHOP_LANG->_PHPSHOP_FILEMANAGER_PUBLISHED => '',
+					$VM_LANG->_PHPSHOP_FILES_LIST_FILENAME => '',
+					$VM_LANG->_PHPSHOP_VIEW => '',
+					$VM_LANG->_PHPSHOP_FILES_LIST_FILETITLE => '',
+					$VM_LANG->_PHPSHOP_UPDATE => '',
+					$VM_LANG->_PHPSHOP_FILES_LIST_FILETYPE => '',
+					$VM_LANG->_PHPSHOP_FILEMANAGER_PUBLISHED => '',
 					_E_REMOVE => "width=\"5%\""
 				);
 $listObj->writeTableHeader( $columns );
@@ -79,8 +84,10 @@ while ($db->next_record()) {
 	// The row number
 	$listObj->addCell( $pageNav->rowNumber( $i ) );
 	
+	$isProductDownload = $db->f("file_title") == $dbf->f("attribute_value") ? true : false;
+	
 	// The Checkbox
-	$listObj->addCell( mosHTML::idBox( $i, $db->f("file_id"), false, "file_id" ) );
+	$listObj->addCell( mosHTML::idBox( $i, $db->f("file_id"), $isProductDownload, "file_id" ) );
 	
 	if($db->f("file_name")) 
 		$tmp_cell = basename($db->f("file_name"));
@@ -95,22 +102,28 @@ while ($db->next_record()) {
 		$thumb = $info["dirname"] ."/resized/". basename($db->f("file_name"),".".$info["extension"])."_".PSHOP_IMG_WIDTH."x".PSHOP_IMG_HEIGHT.".".$info["extension"];
 		$thumburl = str_replace( $mosConfig_absolute_path, $mosConfig_live_site, $thumb );
 		if( is_file( $fullimg ) ) {
-			$tmp_cell = $PHPSHOP_LANG->_PHPSHOP_FILES_LIST_FULL_IMG.":";
-			$tmp_cell .= mm_ToolTip( "&nbsp;<img src=\"".$db->f("file_url") . "\" alt=\"Full Image\" />", $PHPSHOP_LANG->_PHPSHOP_FILES_LIST_FULL_IMG, "[ ".$PHPSHOP_LANG->_PHPSHOP_VIEW . " ]<br/>" ); 
+			$tmp_cell = $VM_LANG->_PHPSHOP_FILES_LIST_FULL_IMG.":";
+			$tmp_cell .= mm_ToolTip( "&nbsp;<img src=\"".$db->f("file_url") . "\" alt=\"Full Image\" />", $VM_LANG->_PHPSHOP_FILES_LIST_FULL_IMG, "[ ".$VM_LANG->_PHPSHOP_VIEW . " ]<br/>" ); 
 		}
 		if( is_file( $thumb ) ) {
-			$tmp_cell .= $PHPSHOP_LANG->_PHPSHOP_FILES_LIST_THUMBNAIL_IMG.":";
-			$tmp_cell .= mm_ToolTip( "&nbsp;<img src=\"$thumburl\" alt=\"thumbnail\" />", $PHPSHOP_LANG->_PHPSHOP_FILES_LIST_THUMBNAIL_IMG, "[ ".$PHPSHOP_LANG->_PHPSHOP_VIEW . " ]" ); 
+			$tmp_cell .= $VM_LANG->_PHPSHOP_FILES_LIST_THUMBNAIL_IMG.":";
+			$tmp_cell .= mm_ToolTip( "&nbsp;<img src=\"$thumburl\" alt=\"thumbnail\" />", $VM_LANG->_PHPSHOP_FILES_LIST_THUMBNAIL_IMG, "[ ".$VM_LANG->_PHPSHOP_VIEW . " ]" ); 
 		}
 		if( !$db->f("file_name") ) {
-			$tmp_cell = "&nbsp;<a target=\"_blank\" href=\"".$db->f("file_url"). "\">[ ".$PHPSHOP_LANG->_PHPSHOP_VIEW . " ]</a><br/>"; 
+			$tmp_cell = "&nbsp;<a target=\"_blank\" href=\"".$db->f("file_url"). "\">[ ".$VM_LANG->_PHPSHOP_VIEW . " ]</a><br/>"; 
 		}
 	}
 	$listObj->addCell( $tmp_cell );
 	
 	$listObj->addCell( $db->f("file_title"));
 	   
-	$tmp_cell = "<a href=\"". $_SERVER['PHP_SELF']."?option=com_phpshop&page=product.file_form&product_id=$product_id&file_id=".$db->f("file_id") ."\">". $PHPSHOP_LANG->_PHPSHOP_FILES_LIST_EDITFILE ."</a>&nbsp;";
+	if( !$isProductDownload )
+		$tmp_cell = "<a href=\"". $_SERVER['PHP_SELF']."?option=com_virtuemart&page=product.file_form&product_id=$product_id&file_id=".$db->f("file_id") ."\">"
+				. $VM_LANG->_PHPSHOP_FILES_LIST_EDITFILE 
+				."</a>&nbsp;";
+	else
+		$tmp_cell = " - ";
+		
     $listObj->addCell( $tmp_cell );	  
 	
 
@@ -125,9 +138,11 @@ while ($db->next_record()) {
 		$tmp_cell = '<img src="'. $mosConfig_live_site .'/administrator/images/tick.png" border="0" alt="Unpublish" />';
 	} 
 	$listObj->addCell( $tmp_cell );
-
-	$listObj->addCell( $ps_html->deleteButton( "file_id", $db->f("file_id"), "deleteProductFile", $keyword, $limitstart, "&product_id=$product_id" ) );
-
+	if( !$isProductDownload )
+		$listObj->addCell( $ps_html->deleteButton( "file_id", $db->f("file_id"), "deleteProductFile", $keyword, $limitstart, "&product_id=$product_id" ) );
+	else
+		$listObj->addCell( "" );
+		
 	$i++;
 
 }
